requirements.txt
pandas>=1.0
scikit-learn>=1.0
joblib
nltk
pytest
pyyaml
streamlit
matplotlib
seaborn